Topology.Node.C: added method 'typ'
'typ' returns the type of a node (Sink, Crossing, etc.)
This is a paradigm shift:
From now on every node is bound to its type.
This way we can assert consistency of types of all nodes in a sequence flow graph.
It saves us from passing the node type through the solver to the draw routines.